requirements.txt
langchain>=0.3.9
langchain_community>=0.3.9
langchain-core>=0.3.21
langgraph>=0.0.11
langchain-aws>=0.2.9
boto3
pandas
numpy
streamlit
python-dotenv
redshift-connector
opensearch-py
pydantic>=2.0.0
pytest
sqlparse

# Opensearch node -> Packages > Add Nori Plugin